@@ -10,16 +10,16 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "~13.0.0",
+    "@angular/animations": "~14.0.3",
     "@angular/cdk": "^13.3.9",
-    "@angular/common": "~13.0.0",
-    "@angular/compiler": "~13.0.0",
-    "@angular/core": "~13.0.0",
-    "@angular/forms": "~13.0.0",
+    "@angular/common": "~14.0.3",
+    "@angular/compiler": "~14.0.3",
+    "@angular/core": "~14.0.3",
+    "@angular/forms": "~14.0.3",
     "@angular/material": "^13.3.9",
-    "@angular/platform-browser": "~13.0.0",
-    "@angular/platform-browser-dynamic": "~13.0.0",
-    "@angular/router": "~13.0.0",
+    "@angular/platform-browser": "~14.0.3",
+    "@angular/platform-browser-dynamic": "~14.0.3",
+    "@angular/router": "~14.0.3",
     "font-awesome": "^4.7.0",
     "rxjs": "~7.4.0",
     "tslib": "^2.3.0",
@@ -27,8 +27,8 @@
   },
   "devDependencies": {
     "@angular-devkit/build-angular": "^14.0.3",
-    "@angular/cli": "^14.0.2",
-    "@angular/compiler-cli": "~13.0.0",
+    "@angular/cli": "^14.0.3",
+    "@angular/compiler-cli": "~14.0.3",
     "@faker-js/faker": "^7.3.0",
     "@types/jasmine": "~3.10.0",
     "@types/node": "^12.11.1",
@@ -39,6 +39,6 @@
     "karma-coverage": "~2.0.3",
     "karma-jasmine": "~4.0.0",
     "karma-jasmine-html-reporter": "~1.7.0",
-    "typescript": "~4.4.3"
+    "typescript": "~4.7.4"
   }
 }
