@@ -1,9 +1,9 @@
 import { Component, Input } from '@angular/core';
-import { FormControl, FormGroupDirective, NgForm } from '@angular/forms';
+import { UntypedFormControl, FormGroupDirective, NgForm } from '@angular/forms';
 import { ErrorStateMatcher } from '@angular/material/core';
 
 export class controlErrorStateMatcher implements ErrorStateMatcher {
-  isErrorState(control: FormControl | null, form: FormGroupDirective | NgForm | null): boolean {
+  isErrorState(control: UntypedFormControl | null, form: FormGroupDirective | NgForm | null): boolean {
     const isSubmitted = form && form.submitted;
     return !!(control && control.invalid && (control.dirty || control.touched || isSubmitted));
   }
@@ -17,7 +17,7 @@ export class controlErrorStateMatcher implements ErrorStateMatcher {
 export class FieldErrorsComponent {
 
   @Input()
-  public formField!: FormControl;
+  public formField!: UntypedFormControl;
 
   @Input()
   public fieldName!: string;
