@@ -2,7 +2,7 @@ import { PostApiService as MockPostApiService } from 'src/app/modules/posts/serv
 import { PostApiService } from 'src/app/modules/posts/services/api/posts-api.service';
 import { CommentApiService as MockCommentApiService } from 'src/app/modules/comments/services/api/comment-api.mock.service';
 import { CommentApiService } from 'src/app/modules/comments/services/api/comment-api.service';
-import { PostApiService as MockManagePostApiService } from 'src/app/modules/posts/services/api/posts-api.mock.service';
+import { ManagePostApiService as MockManagePostApiService } from 'src/app/modules/manage-posts/services/api/manage-post-api.mock.service';
 import { ManagePostApiService } from 'src/app/modules/manage-posts/services/api/manage-post-api.service';
 
 export const environment = {
